INCORRECT. Still some failing benches to check out! Changes to PBSolver: 	replaced long long with int64_t and analogous for unsigned version. 	took care of [0] initialized arrays and effectively missing includes. 	Could NOT fix arrays initialized with variables, which is also not allowed in iso C++ -> this is only remaining issue with iso c++ errors.
Rewritten argument parsing using tclap and in new files. Not really code reduction, but certainly more readable and maintainable. Putting the arguments in the separate classes was not really usefull as the global option class which is passed around needs to have knowledge of them anyway and has to be initialized explicitly in main. Renamed ECNF_mode to SolverOption.

Moved checking for well-foundedness of definition to definition solver, adding generic checkStatus above it.

+++ pbsolver/pbbase/h/BaseSearchState.h
@@ -1,15 +1,17 @@
 #ifndef BASESEARCHSTATE_H_
 #define BASESEARCHSTATE_H_
 
+#include "Global.h"
+
 namespace MiniSatPP {
 struct BaseSearchState {
     BaseSearchState* parent;
     int index;
 	int lastRelevent;
 	unsigned int baseMul;
-	unsigned long long cost;
-	unsigned long long hCost;
-	unsigned long long carryins;
+	uint64 cost;
+	uint64 hCost;
+	uint64 carryins;
 	BaseSearchState(BaseSearchState* parent_, int lastRelevent_ ,unsigned int baseMul_,
 					unsigned long cost_,unsigned long hCost_,
 					unsigned long carryins_) :
