+++ pbsolver/pbbase/h/SearchMetaData.h
@@ -6,18 +6,20 @@
 #include <vector>
 #include <string>
 
+#include "Global.h"
+
 namespace MiniSatPP {
 class SearchMetaData {
 	
   public:
   	std::vector<int> base;
-  	std::vector<unsigned long long> carry;
-  	std::vector<unsigned long long> inputs;
-  	unsigned long long emptyBaseNOI;
-	unsigned long long cost;
-	unsigned long long inputCountCost;
-	unsigned long long carryOnlyCost;
-	unsigned long long compCost;
+  	std::vector<uint64> carry;
+  	std::vector<uint64> inputs;
+  	uint64 emptyBaseNOI;
+	uint64 cost;
+	uint64 inputCountCost;
+	uint64 carryOnlyCost;
+	uint64 compCost;
 	unsigned long basesEvaluated;
 	unsigned long runTime; //run time in micro secounds
 	unsigned long insertedToQue;
@@ -30,7 +32,7 @@ class SearchMetaData {
 	std::string algType;
 	
 	SearchMetaData(int size,unsigned int primesCutOf_,unsigned int mxcoff,int len,std::string algt);
-	void finalize(unsigned long long cost);
+	void finalize(uint64 cost);
 	virtual ~SearchMetaData();
 	void print();
